NinelTour_12/requirements.txt
adjacent==1.1.0
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
cent==4.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-elasticsearch-dsl==7.2.1
django-sslserver==0.22
django-timezone-field==4.2.1
djangorestframework==3.12.4
elasticsearch==7.16.0
elasticsearch-dsl==7.4.0
flower==1.0.0
future==0.18.2
humanize==3.13.1
idna==3.3
kombu==5.2.2
oauthlib==3.1.1
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
serializers==0.2.4
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
